# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "DecisionBatchAgent": {
        "additionalProperties": false,
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "trade_count": {
            "title": "Trade Count",
            "type": "integer"
          }
        },
        "required": [
          "status",
          "completed_at",
          "error",
          "trade_count"
        ],
        "title": "DecisionBatchAgent",
        "type": "object"
      },
      "DecisionBatchCounts": {
        "additionalProperties": false,
        "properties": {
          "completed": {
            "title": "Completed",
            "type": "integer"
          },
          "failed": {
            "title": "Failed",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "completed",
          "failed"
        ],
        "title": "DecisionBatchCounts",
        "type": "object"
      },
      "DecisionBatchStatus": {
        "additionalProperties": false,
        "properties": {
          "agents": {
            "additionalProperties": {
              "$ref": "#/components/schemas/DecisionBatchAgent"
            },
            "title": "Agents",
            "type": "object"
          },
          "batch_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch Id"
          },
          "counts": {
            "$ref": "#/components/schemas/DecisionBatchCounts"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "last_completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Completed At"
          },
          "last_triggered_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Triggered At"
          },
          "next_eligible_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Eligible At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "batch_id",
          "status",
          "last_triggered_at",
          "last_completed_at",
          "next_eligible_at",
          "counts",
          "agents"
        ],
        "title": "DecisionBatchStatus",
        "type": "object"
      },
      "DecisionDay": {
        "additionalProperties": false,
        "properties": {
          "batch": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DecisionBatchStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "date": {
            "title": "Date",
            "type": "string"
          },
          "due_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due At"
          },
          "is_today": {
            "title": "Is Today",
            "type": "boolean"
          },
          "run_count": {
            "title": "Run Count",
            "type": "integer"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "weekday": {
            "title": "Weekday",
            "type": "string"
          }
        },
        "required": [
          "date",
          "weekday",
          "is_today",
          "state",
          "due_at",
          "batch",
          "run_count"
        ],
        "title": "DecisionDay",
        "type": "object"
      },
      "DecisionSchedule": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "time": {
            "title": "Time",
            "type": "string"
          },
          "weekdays": {
            "items": {
              "type": "integer"
            },
            "title": "Weekdays",
            "type": "array"
          }
        },
        "required": [
          "kind",
          "weekdays",
          "time"
        ],
        "title": "DecisionSchedule",
        "type": "object"
      },
      "DecisionWeekResponse": {
        "additionalProperties": false,
        "properties": {
          "ai_account_count": {
            "title": "Ai Account Count",
            "type": "integer"
          },
          "current_batch": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DecisionBatchStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "days": {
            "items": {
              "$ref": "#/components/schemas/DecisionDay"
            },
            "title": "Days",
            "type": "array"
          },
          "latest_batch": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DecisionBatchStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "next_reminder_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Reminder At"
          },
          "schedule": {
            "$ref": "#/components/schemas/DecisionSchedule"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          },
          "week_start": {
            "title": "Week Start",
            "type": "string"
          }
        },
        "required": [
          "week_start",
          "timezone",
          "schedule",
          "days",
          "current_batch",
          "latest_batch",
          "next_reminder_at",
          "ai_account_count"
        ],
        "title": "DecisionWeekResponse",
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "items": {
                  "$ref": "#/components/schemas/ValidationIssue"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "error": {
            "title": "Error",
            "type": "string"
          },
          "ok": {
            "const": false,
            "default": false,
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "error",
          "code"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HoldingResponse": {
        "additionalProperties": false,
        "properties": {
          "average_cost": {
            "title": "Average Cost",
            "type": "number"
          },
          "current_price": {
            "title": "Current Price",
            "type": "number"
          },
          "market_value": {
            "title": "Market Value",
            "type": "number"
          },
          "opened_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opened At"
          },
          "pnl": {
            "title": "Pnl",
            "type": "number"
          },
          "pnl_percent": {
            "title": "Pnl Percent",
            "type": "number"
          },
          "quantity": {
            "title": "Quantity",
            "type": "number"
          },
          "ticker": {
            "title": "Ticker",
            "type": "string"
          }
        },
        "required": [
          "ticker",
          "opened_at",
          "quantity",
          "average_cost",
          "current_price",
          "market_value",
          "pnl",
          "pnl_percent"
        ],
        "title": "HoldingResponse",
        "type": "object"
      },
...
      "LeaderboardEntry": {
        "additionalProperties": false,
        "properties": {
          "cash_balance": {
            "title": "Cash Balance",
            "type": "number"
          },
          "decision_architecture": {
            "title": "Decision Architecture",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "holdings": {
            "items": {
              "$ref": "#/components/schemas/HoldingResponse"
            },
            "title": "Holdings",
            "type": "array"
          },
          "holdings_count": {
            "title": "Holdings Count",
            "type": "integer"
          },
          "holdings_value": {
            "title": "Holdings Value",
            "type": "number"
          },
          "pnl_percent": {
            "title": "Pnl Percent",
            "type": "number"
          },
          "pnl_total": {
            "title": "Pnl Total",
            "type": "number"
          },
          "rank": {
            "title": "Rank",
            "type": "integer"
          },
          "realized_pnl": {
            "title": "Realized Pnl",
            "type": "number"
          },
          "total_value": {
            "title": "Total Value",
            "type": "number"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          },
          "user_type": {
            "title": "User Type",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "username",
          "display_name",
          "user_type",
          "decision_architecture",
          "cash_balance",
          "holdings_value",
          "total_value",
          "pnl_total",
          "pnl_percent",
          "realized_pnl",
          "holdings",
          "holdings_count",
          "rank"
        ],
        "title": "LeaderboardEntry",
        "type": "object"
      },
...
      "NewsItemResponse": {
        "additionalProperties": false,
        "properties": {
          "canonical_url": {
            "title": "Canonical Url",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "published_at": {
            "title": "Published At",
            "type": "string"
          },
          "publisher": {
            "title": "Publisher",
            "type": "string"
          },
          "source_tier": {
            "title": "Source Tier",
            "type": "integer"
          },
          "ticker": {
            "title": "Ticker",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "ticker",
          "id",
          "title",
          "publisher",
          "provider",
          "canonical_url",
          "published_at",
          "source_tier"
        ],
        "title": "NewsItemResponse",
        "type": "object"
      },
...
      "PerformanceResponse": {
        "additionalProperties": false,
        "properties": {
          "buys": {
            "title": "Buys",
            "type": "integer"
          },
          "cash": {
            "title": "Cash",
            "type": "number"
          },
          "decision_architecture": {
            "title": "Decision Architecture",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "pnl_percent": {
            "title": "Pnl Percent",
            "type": "number"
          },
          "pnl_total": {
            "title": "Pnl Total",
            "type": "number"
          },
          "portfolio_value": {
            "title": "Portfolio Value",
            "type": "number"
          },
          "positions": {
            "title": "Positions",
            "type": "integer"
          },
          "sells": {
            "title": "Sells",
            "type": "integer"
          },
          "total_bought": {
            "title": "Total Bought",
            "type": "number"
          },
          "total_sold": {
            "title": "Total Sold",
            "type": "number"
          },
          "total_trades": {
            "title": "Total Trades",
            "type": "integer"
          },
          "user_type": {
            "title": "User Type",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "display_name",
          "user_type",
          "decision_architecture",
          "portfolio_value",
          "cash",
          "pnl_total",
          "pnl_percent",
          "total_trades",
          "buys",
          "sells",
          "total_bought",
          "total_sold",
          "positions"
        ],
        "title": "PerformanceResponse",
        "type": "object"
      },
...
      "PortfolioHistoryResponse": {
        "additionalProperties": false,
        "properties": {
          "history": {
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/HistoryPoint"
              },
              "type": "array"
            },
            "title": "History",
            "type": "object"
          },
          "users": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Users",
            "type": "object"
          }
        },
        "required": [
          "history",
          "users"
        ],
        "title": "PortfolioHistoryResponse",
        "type": "object"
      },
...
      "TransactionResponse": {
        "additionalProperties": false,
        "properties": {
          "cash_balance_after": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cash Balance After"
          },
          "cash_balance_before": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cash Balance Before"
          },
          "decision_snapshot_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision Snapshot At"
          },
          "executed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Executed At"
          },
          "execution_market_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution Market State"
          },
          "execution_quote_audit_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution Quote Audit Id"
          },
          "execution_quote_captured_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution Quote Captured At"
          },
          "execution_quote_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution Quote Source"
          },
          "funnel_cycle_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Funnel Cycle Id"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "llm_reasoning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Reasoning"
          },
          "market_closed": {
            "title": "Market Closed",
            "type": "integer"
          },
          "price_per_share": {
            "title": "Price Per Share",
            "type": "number"
          },
          "quantity": {
            "title": "Quantity",
            "type": "number"
          },
          "realized_pnl": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Realized Pnl"
          },
          "ticker": {
            "title": "Ticker",
            "type": "string"
          },
          "total_value": {
            "title": "Total Value",
            "type": "number"
          },
          "transaction_type": {
            "title": "Transaction Type",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          }
        },
        "required": [
          "id",
          "user_id",
          "ticker",
          "transaction_type",
          "quantity",
          "price_per_share",
          "total_value",
          "cash_balance_before",
          "cash_balance_after",
          "llm_reasoning",
          "funnel_cycle_id",
          "market_closed",
          "realized_pnl",
          "execution_quote_audit_id",
          "executed_at"
        ],
        "title": "TransactionResponse",
        "type": "object"
      },
      "ValidationIssue": {
        "additionalProperties": false,
        "properties": {
          "location": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "location",
          "message",
          "type"
        ],
        "title": "ValidationIssue",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Portfolio Simulator",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/decision-batches": {
      "post": {
        "operationId": "create_api_decision_batches_post",
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionBatchStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Operator access required"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Operation conflict"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Create",
        "tags": [
          "decisions"
        ]
      }
    },
    "/api/decision-batches/status": {
      "get": {
        "operationId": "status_api_decision_batches_status_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionBatchStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Status",
        "tags": [
          "decisions"
        ]
      }
    },
    "/api/decision-batches/week": {
      "get": {
        "operationId": "week_status_api_decision_batches_week_get",
        "parameters": [
          {
            "in": "query",
            "name": "week_start",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Week Start"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionWeekResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Week Status",
        "tags": [
          "decisions"
        ]
      }
    },
    "/api/export/csv": {
      "get": {
        "description": "Export all transactions as CSV.",
        "operationId": "export_csv_api_export_csv_get",
        "responses": {
          "200": {
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Export Csv",
        "tags": [
          "dashboard"
        ]
      }
    },
...
    "/api/leaderboard": {
      "get": {
        "operationId": "leaderboard_api_leaderboard_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LeaderboardEntry"
                  },
                  "title": "Response Leaderboard Api Leaderboard Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Leaderboard",
        "tags": [
          "dashboard"
        ]
      }
    },
    "/api/news": {
      "get": {
        "operationId": "news_api_news_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 12,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NewsItemResponse"
                  },
                  "title": "Response News Api News Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "News",
        "tags": [
          "dashboard"
        ]
      }
    },
...
    "/api/portfolio-history": {
      "get": {
        "operationId": "portfolio_history_api_portfolio_history_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioHistoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Portfolio History",
        "tags": [
          "dashboard"
        ]
      }
    },
...
    "/api/stats": {
      "get": {
        "operationId": "performance_stats_api_stats_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PerformanceResponse"
                  },
                  "title": "Response Performance Stats Api Stats Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Performance Stats",
        "tags": [
          "dashboard"
        ]
      }
    },
...
    "/api/transactions": {
      "get": {
        "operationId": "transactions_api_transactions_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 30,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TransactionResponse"
                  },
                  "title": "Response Transactions Api Transactions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Transactions",
        "tags": [
          "dashboard"
        ]
      }
    },
...
  }
}